requirements.txt
asgiref==3.3.1
Django==3.1.6
django-ckeditor==6.0.0
django-environ==0.4.5
django-js-asset==1.2.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
